feat: our own scalacheck implicit removed
